Grouped up files into packages, full path is now logged while downloading a file, renamed YtDownloadFile class and made it hold videoUrlPart
