+++ src/main/java/machalica/marcin/spring/ytdownloader/downloader/YtDownloaderDao.java
@@ -4,7 +4,9 @@ import java.io.File;
 
 import com.sapher.youtubedl.YoutubeDLException;
 
+import machalica.marcin.spring.ytdownloader.helpers.YtFileInfo;
+
 public interface YtDownloaderDao {
-	YtDownloadFile getFileInfo(String videoUrl) throws YoutubeDLException;
+	YtFileInfo getFileInfo(String videoUrl) throws YoutubeDLException;
 	File downloadFile(String videoUrl, String format) throws YoutubeDLException;
 }
