+++ src/main/java/machalica/marcin/spring/ytdownloader/qrcodegenerator/QrCodeGenerator.java
@@ -1,4 +1,4 @@
-package machalica.marcin.spring.ytdownloader.downloader;
+package machalica.marcin.spring.ytdownloader.qrcodegenerator;
 
 import java.io.ByteArrayOutputStream;
 import java.net.InetAddress;
